[FIX] : 실제 categoryVideo list 데이터 따로 보관
